refacto update company details and business type
