+++ web/admin/modals/UpdateCompanyDetailsModal.js
@@ -1,7 +1,5 @@
 import React from "react";
 
-import { useApi } from "common/utils/api";
-
 import {
   Button,
   Checkbox,
@@ -13,14 +11,10 @@ import {
 } from "@dataesr/react-dsfr";
 import { makeStyles } from "@mui/styles";
 import { PhoneNumber } from "../../common/PhoneNumber";
-import {
-  UPDATE_COMPANY_DETAILS,
-  UPDATE_COMPANY_DETAILS_WITH_BUSINESS_TYPE
-} from "common/utils/apiQueries";
-import { ADMIN_ACTIONS } from "../store/reducers/root";
+
 import Stack from "@mui/material/Stack";
-import { useSnackbarAlerts } from "../../common/Snackbar";
 import { BusinessType } from "../../common/BusinessType";
+import { useUpdateCompanyDetails } from "../../common/useUpdateCompanyDetails";
 
 const useStyles = makeStyles(theme => ({
   label: {
@@ -35,25 +29,20 @@ export default function UpdateCompanyDetailsModal({
   adminStore
 }) {
   const classes = useStyles();
-  const api = useApi();
-  const alerts = useSnackbarAlerts();
+  const {
+    newCompanyName,
+    setNewCompanyName,
+    newCompanyPhoneNumber,
+    setNewCompanyPhoneNumber,
+    setNewCompanyBusinessType,
+    hasBusinessTypeChanged,
+    updateCompanyDetails
+  } = useUpdateCompanyDetails(company, adminStore, handleClose);
 
-  const [newCompanyName, setNewCompanyName] = React.useState(company?.name);
-  const [newCompanyPhoneNumber, setNewCompanyPhoneNumber] = React.useState(
-    company?.phoneNumber
-  );
-  const [newCompanyBusinessType, setNewCompanyBusinessType] = React.useState(
-    adminStore.business?.BusinessType
-  );
   const [
     applyBusinessTypeToEmployees,
     setApplyBusinessTypeToEmployees
   ] = React.useState(false);
-
-  const hasBusinessTypeChanged = React.useMemo(
-    () => newCompanyBusinessType !== adminStore.business?.businessType,
-    [newCompanyBusinessType, adminStore.business?.businessType]
-  );
 
   const canSave = React.useMemo(
     () =>
@@ -70,66 +59,9 @@ export default function UpdateCompanyDetailsModal({
     ]
   );
 
-  const handleSubmit = async () => {
-    await alerts.withApiErrorHandling(async () => {
-      const apiResponse = await api.graphQlMutate(
-        hasBusinessTypeChanged
-          ? UPDATE_COMPANY_DETAILS_WITH_BUSINESS_TYPE
-          : UPDATE_COMPANY_DETAILS,
-        {
-          companyId: company?.id,
-          newName: newCompanyName,
-          newPhoneNumber: newCompanyPhoneNumber,
-          ...(hasBusinessTypeChanged
-            ? {
-                newBusinessType: newCompanyBusinessType,
-                applyBusinessTypeToEmployees
-              }
-            : {})
-        },
-        { context: { nonPublicApi: false } }
-      );
+  const handleSubmit = async () =>
+    await updateCompanyDetails(applyBusinessTypeToEmployees);
 
-      const { id, business } = apiResponse?.data?.updateCompanyDetails;
-      await adminStore.dispatch({
-        type: ADMIN_ACTIONS.updateCompanyNameAndPhoneNumber,
-        payload: {
-          companyId: id,
-          companyName: newCompanyName,
-          companyPhoneNumber: newCompanyPhoneNumber
-        }
-      });
-      await adminStore.dispatch({
-        type: ADMIN_ACTIONS.updateBusinessType,
-        payload: {
-          business
-        }
-      });
-      if (hasBusinessTypeChanged) {
-        const { employments } = apiResponse?.data?.updateCompanyDetails;
-        for (const employment of employments) {
-          await adminStore.dispatch({
-            type: ADMIN_ACTIONS.update,
-            payload: {
-              id: employment.id,
-              entity: "employments",
-              update: {
-                ...employment,
-                companyId: id,
-                adminStore
-              }
-            }
-          });
-        }
-      }
-      alerts.success(
-        "Les détails de l'entreprise ont bien été enregistrés",
-        "",
-        6000
-      );
-      handleClose();
-    }, "update-company-details");
-  };
   return (
     <Modal isOpen={open} hide={handleClose} size="lg">
       <ModalTitle>Modifier les détails de l'entreprise</ModalTitle>
